+++ src/structlog/processors.py
@@ -30,13 +30,15 @@ class KeyValueRenderer(object):
 
     :param bool sort_keys: Whether to sort keys when formatting.
     :param list key_order: List of keys that should be rendered in this exact
-        order.  Missing keys will be rendered as `None`, extra keys depending
-        on *sort_keys* and the dict class unless drop_missing is True.
-    :param bool drop_missing: When True, extra keys in key_order will be
-        dropped rather than rendered as None
+        order.  Missing keys will be rendered as ``None``, extra keys depending
+        on *sort_keys* and the dict class.
+    :param bool drop_missing: When True, extra keys in *key_order* will be
+        dropped rather than rendered as ``None``.
 
     .. versionadded:: 0.2.0
-        `key_order`
+        *key_order*
+    .. versionadded:: 16.1.0
+        *drop_missing*
     """
     def __init__(self, sort_keys=False, key_order=None, drop_missing=False):
         # Use an optimized version for each case.